+++ crates/import/src/lib.rs
@@ -21,37 +21,55 @@ pub use insomnia::from_insomnia;
 pub use openapi::from_openapi;
 use reqwest::{Response, Url};
 pub use rest::from_rest;
-use tokio::fs;
+use tokio::{
+    fs,
+    io::{self, AsyncReadExt, BufReader},
+};
 use tracing::info;
 
 /// Pointer to a file that should be imported
 #[derive(Clone, Debug, derive_more::Display)]
 pub enum ImportInput {
-    #[display("{_0}")]
-    Url(Url),
+    /// Import data from stdin
+    #[display("stdin")]
+    Stdin,
+    /// Import from a local file
     #[display("{}", _0.display())]
     Path(PathBuf),
+    /// Download a file via HTTP and import it
+    #[display("{_0}")]
+    Url(Url),
 }
 
 impl ImportInput {
     /// Load the contents of the import input
     async fn load(&self) -> anyhow::Result<String> {
         match self {
-            Self::Url(url) => {
-                info!(%url, "Fetching remote file for import");
-                let content = reqwest::get(url.clone())
-                    .and_then(Response::text)
+            Self::Stdin => {
+                info!("Reading stdin for import");
+                let mut reader = BufReader::new(io::stdin());
+                let mut content = String::with_capacity(1024);
+                reader
+                    .read_to_string(&mut content)
                     .await
-                    .with_context(|| {
-                        format!("Error importing HTTP URL {url}")
-                    })?;
+                    .context("Error importing from stdin")?;
                 Ok(content)
             }
             Self::Path(path) => {
                 info!(?path, "Reading local file for import");
                 let content =
                     fs::read_to_string(path).await.with_context(|| {
-                        format!("Error importing local file {path:?}")
+                        format!("Error importing from local file {path:?}")
+                    })?;
+                Ok(content)
+            }
+            Self::Url(url) => {
+                info!(%url, "Fetching remote file for import");
+                let content = reqwest::get(url.clone())
+                    .and_then(Response::text)
+                    .await
+                    .with_context(|| {
+                        format!("Error importing from HTTP URL {url}")
                     })?;
                 Ok(content)
             }
@@ -60,30 +78,12 @@ impl ImportInput {
 
     /// Get the name of the input file. For a path this just grabs the
     /// file name from the path. For a URL it treats the path component of the
-    /// URL as a file path and gets the file namefrom there.
-    ///
-    /// ```
-    /// assert_eq!(
-    ///     ImportInput::from_str("./openapi.json").unwrap().file_name(),
-    ///     Some("openapi.json")
-    /// );
-    /// assert_eq!(
-    ///     ImportInput::from_str("https://example.com/openapi.json")
-    ///         .unwrap()
-    ///         .file_name(),
-    ///     Some("openapi.json")
-    /// );
-    /// assert_eq!(
-    ///     ImportInput::from_str("https://example.com/")
-    ///         .unwrap()
-    ///         .file_name(),
-    ///     None
-    /// );
-    /// ```
+    /// URL as a file path and gets the file namefrom there
     fn file_name(&self) -> Option<&OsStr> {
         match self {
-            Self::Url(url) => Path::new(url.path()).file_name(),
+            Self::Stdin => None,
             Self::Path(path) => path.file_name(),
+            Self::Url(url) => Path::new(url.path()).file_name(),
         }
     }
 }
@@ -92,9 +92,17 @@ impl FromStr for ImportInput {
     type Err = Infallible;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        // If it's a URL, it's a URL. Otherwise, it's a path
-        if let Ok(url) = s.parse::<Url>() {
-            Ok(Self::Url(url))
+        if s == "-" {
+            Ok(Self::Stdin)
+        } else if let Ok(url) = s.parse::<Url>() {
+            // Windows paths (C:\...) parse as URLs, so we need to make sure
+            // it's an HTTP URL
+            // unstable: if-let chain (Rust 1.88)
+            if ["http", "https"].contains(&url.scheme()) {
+                Ok(Self::Url(url))
+            } else {
+                Ok(Self::Path(PathBuf::from(s)))
+            }
         } else {
             Ok(Self::Path(PathBuf::from(s)))
         }